Create bins creation
function for creating the bins and normalizing them, the function accepts the table of measurements from the experiment (i.e. the time intervals) and a range, and return a matrix whos columns are:
x=delta_t
dx=the error in delta_t
y=the number of events in the bin
dy=the error in the number of events
the bins are determined by the range.